requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
aiomysql==0.2.0
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
greenlet==3.1.1
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
aiosqlite==0.20.0
boto3==1.36.26
# ultralytics es opcional para inferencia cloud real de inventario_foto.
# ultralytics==8.3.74
